transitions v2

+++ src/scenes/empty_scene.h
@@ -1,16 +1,12 @@
-#include "scene.h"
 #include "app_context.h"
+#include "scene.h"
 
 #define THIS_SCENE EmptyScene
 
-void THIS_SCENE::Enter(AppContext& ctx) {
-}
+void THIS_SCENE::Enter(AppContext& ctx) {}
 
-void THIS_SCENE::Leave(AppContext& ctx) {
-}
+void THIS_SCENE::Leave(AppContext& ctx) {}
 
-void THIS_SCENE::Update(AppContext& ctx, float dt) {
-}
+void THIS_SCENE::Update(AppContext& ctx, double dt) {}
 
-void THIS_SCENE::Draw(AppContext& ctx) {
-}
+void THIS_SCENE::Draw(AppContext& ctx) {}
