+++ src/scenes/main_menu_scene.cpp
@@ -12,7 +12,7 @@ void THIS_SCENE::Enter(AppContext& ctx) {
 
 void THIS_SCENE::Leave(AppContext& ctx) {}
 
-void THIS_SCENE::Update(AppContext& ctx, float dt) {
+void THIS_SCENE::Update(AppContext& ctx, double dt) {
     if (ctx.pad.IsPressed(OrbisPadButton::ORBIS_PAD_BUTTON_CIRCLE)) {
         ctx.running = false;
     }
